+++ core-ng/src/main/java/core/framework/impl/scheduler/FixedRateTask.java
@@ -2,19 +2,18 @@ package core.framework.impl.scheduler;
 
 
 import core.framework.scheduler.Job;
-import core.framework.util.Randoms;
 
 import java.time.Duration;
 
 /**
  * @author neo
  */
-public final class FixedRateTrigger implements Trigger {
+public final class FixedRateTask implements Task {
+    final Duration rate;
     private final String name;
     private final Job job;
-    private final Duration rate;
 
-    public FixedRateTrigger(String name, Job job, Duration rate) {
+    FixedRateTask(String name, Job job, Duration rate) {
         this.name = name;
         this.job = job;
         this.rate = rate;
@@ -31,13 +30,7 @@ public final class FixedRateTrigger implements Trigger {
     }
 
     @Override
-    public void schedule(Scheduler scheduler) {
-        Duration delay = Duration.ofMillis((long) Randoms.number(1000, 3000)); // delay 1s to 3s
-        scheduler.schedule(this, delay, rate);
-    }
-
-    @Override
-    public String frequency() {
+    public String trigger() {
         return "fixedRate@" + rate;
     }
 }