+++ core-ng/src/main/java/core/framework/impl/scheduler/Scheduler.java
@@ -4,13 +4,16 @@ import core.framework.impl.async.ThreadPools;
 import core.framework.impl.log.ActionLog;
 import core.framework.impl.log.LogManager;
 import core.framework.scheduler.Job;
+import core.framework.scheduler.Trigger;
 import core.framework.util.Exceptions;
 import core.framework.util.Maps;
+import core.framework.util.Randoms;
 import core.framework.web.exception.NotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.time.Duration;
+import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
@@ -21,20 +24,31 @@ import java.util.concurrent.TimeUnit;
  * @author neo
  */
 public final class Scheduler {
-    public final Map<String, Trigger> triggers = Maps.newHashMap();
+    public final Map<String, Task> tasks = Maps.newHashMap();
     private final Logger logger = LoggerFactory.getLogger(Scheduler.class);
     private final ScheduledExecutorService scheduler = ThreadPools.singleThreadScheduler("scheduler-");
     private final ExecutorService jobExecutor = ThreadPools.cachedThreadPool(Runtime.getRuntime().availableProcessors() * 4, "scheduler-job-");
     private final LogManager logManager;
+    public ZoneId zoneId = ZoneId.systemDefault();
 
     public Scheduler(LogManager logManager) {
         this.logManager = logManager;
     }
 
     public void start() {
-        triggers.forEach((name, trigger) -> {
-            logger.info("schedule job, job={}, frequency={}, jobClass={}", name, trigger.frequency(), trigger.job().getClass().getCanonicalName());
-            trigger.schedule(this);
+        ZonedDateTime now = ZonedDateTime.now(zoneId);
+        tasks.forEach((name, task) -> {
+            logger.info("schedule job, job={}, trigger={}, zone={}, jobClass={}", name, task.trigger(), zoneId, task.job().getClass().getCanonicalName());
+            if (task instanceof FixedRateTask) {
+                schedule((FixedRateTask) task);
+            } else if (task instanceof TriggerTask) {
+                try {
+                    ZonedDateTime next = next(((TriggerTask) task).trigger, now);
+                    schedule((TriggerTask) task, next);
+                } catch (Throwable e) {
+                    logger.error("failed to schedule job, job={}", name, e);  // next() with custom trigger impl may throw exception, we don't let runtime error fail startup
+                }
+            }
         });
         logger.info("scheduler started");
     }
@@ -50,41 +64,65 @@ public final class Scheduler {
         }
     }
 
-    public void addTrigger(Trigger trigger) {
-        Class<? extends Job> jobClass = trigger.job().getClass();
+    public void addFixedRateTask(String name, Job job, Duration rate) {
+        addTask(new FixedRateTask(name, job, rate));
+    }
+
+    public void addTriggerTask(String name, Job job, Trigger trigger) {
+        addTask(new TriggerTask(name, job, trigger, zoneId));
+    }
+
+    private void addTask(Task task) {
+        Class<? extends Job> jobClass = task.job().getClass();
         if (jobClass.isSynthetic())
             throw Exceptions.error("job class must not be anonymous class or lambda, please create static class, jobClass={}", jobClass.getCanonicalName());
 
-        String name = trigger.name();
-        Trigger previous = triggers.putIfAbsent(name, trigger);
+        String name = task.name();
+        Task previous = tasks.putIfAbsent(name, task);
         if (previous != null)
             throw Exceptions.error("found duplicate job, name={}, previous={}", name, previous.job().getClass().getCanonicalName());
     }
 
-    void schedule(DynamicTrigger trigger, ZonedDateTime next) {
-        ZonedDateTime now = ZonedDateTime.now();
-        Duration delay = Duration.between(now, next);
-        scheduler.schedule(new DynamicTriggerJob(this, trigger, next), delay.toNanos(), TimeUnit.NANOSECONDS);
+    private ZonedDateTime next(Trigger trigger, ZonedDateTime previous) {
+        ZonedDateTime next = trigger.next(previous);
+        if (next == null || !next.isAfter(previous)) throw Exceptions.error("next scheduled time must be after previous, previous={}, next={}", previous, next);
+        return next;
     }
 
-    void schedule(Trigger trigger, Duration delay, Duration rate) {
-        scheduler.scheduleAtFixedRate(new FixedRateTriggerJob(this, trigger), delay.toNanos(), rate.toNanos(), TimeUnit.NANOSECONDS);
+    private void schedule(TriggerTask task, ZonedDateTime time) {
+        ZonedDateTime now = ZonedDateTime.now(zoneId);
+        Duration delay = Duration.between(now, time);
+        scheduler.schedule(() -> {
+            ZonedDateTime next = next(task.trigger, time);
+            schedule(task, next);
+            logger.info("execute scheduled job, job={}, time={}, next={}", task.name(), time, next);
+            submitJob(task, false);
+        }, delay.toNanos(), TimeUnit.NANOSECONDS);
+    }
+
+    private void schedule(FixedRateTask task) {
+        Duration delay = Duration.ofMillis((long) Randoms.number(1000, 3000)); // delay 1s to 3s
+        scheduler.scheduleAtFixedRate(() -> {
+            logger.info("execute scheduled job, job={}", task.name());
+            submitJob(task, false);
+        }, delay.toNanos(), task.rate.toNanos(), TimeUnit.NANOSECONDS);
     }
 
     public void triggerNow(String name) {
-        Trigger trigger = triggers.get(name);
-        if (trigger == null) throw new NotFoundException("job not found, name=" + name);
-        submitJob(trigger, true);
+        Task task = tasks.get(name);
+        if (task == null) throw new NotFoundException("job not found, name=" + name);
+        submitJob(task, true);
     }
 
-    void submitJob(Trigger trigger, boolean trace) {
+    private void submitJob(Task task, boolean trace) {
         jobExecutor.submit(() -> {
             try {
                 ActionLog actionLog = logManager.begin("=== job execution begin ===");
-                String name = trigger.name();
+                String name = task.name();
                 actionLog.action("job:" + name);
                 actionLog.trace = trace;
-                Job job = trigger.job();
+                actionLog.context("trigger", task.trigger());
+                Job job = task.job();
                 actionLog.context("job", name);
                 actionLog.context("jobClass", job.getClass().getCanonicalName());
                 job.execute();
@@ -97,5 +135,4 @@ public final class Scheduler {
             }
         });
     }
-
 }