+++ core-ng/src/main/java/core/framework/impl/scheduler/Task.java
@@ -5,12 +5,10 @@ import core.framework.scheduler.Job;
 /**
  * @author neo
  */
-public interface Trigger {
+public interface Task {
     String name();
 
     Job job();
 
-    String frequency();
-
-    void schedule(Scheduler scheduler);
+    String trigger();
 }