+++ core-ng/src/main/java/core/framework/module/SchedulerConfig.java
@@ -3,13 +3,12 @@ package core.framework.module;
 import core.framework.http.HTTPMethod;
 import core.framework.impl.module.ModuleContext;
 import core.framework.impl.scheduler.DailyTrigger;
-import core.framework.impl.scheduler.FixedRateTrigger;
 import core.framework.impl.scheduler.MonthlyTrigger;
 import core.framework.impl.scheduler.Scheduler;
-import core.framework.impl.scheduler.SecondlyTrigger;
 import core.framework.impl.scheduler.WeeklyTrigger;
 import core.framework.impl.web.management.SchedulerController;
 import core.framework.scheduler.Job;
+import core.framework.scheduler.Trigger;
 
 import java.time.DayOfWeek;
 import java.time.Duration;
@@ -32,32 +31,28 @@ public final class SchedulerConfig {
     public void timeZone(ZoneId zoneId) {
         if (state.triggerAdded) throw new Error("schedule().timeZone() must be configured before adding trigger");
         if (zoneId == null) throw new Error("zoneId must not be null");
-        state.zoneId = zoneId;
+        state.scheduler.zoneId = zoneId;
     }
 
     public void fixedRate(String name, Job job, Duration rate) {
-        state.scheduler.addTrigger(new FixedRateTrigger(name, job, rate));
-        state.triggerAdded = true;
-    }
-
-    // run every X seconds and align with the closest exact start of minute
-    public void secondly(String name, Job job, int rateInSeconds) {
-        state.scheduler.addTrigger(new SecondlyTrigger(name, job, rateInSeconds));
+        state.scheduler.addFixedRateTask(name, job, rate);
         state.triggerAdded = true;
     }
 
     public void dailyAt(String name, Job job, LocalTime time) {
-        state.scheduler.addTrigger(new DailyTrigger(name, job, time, state.zoneId));
-        state.triggerAdded = true;
+        trigger(name, job, new DailyTrigger(time));
     }
 
     public void weeklyAt(String name, Job job, DayOfWeek dayOfWeek, LocalTime time) {
-        state.scheduler.addTrigger(new WeeklyTrigger(name, job, dayOfWeek, time, state.zoneId));
-        state.triggerAdded = true;
+        trigger(name, job, new WeeklyTrigger(dayOfWeek, time));
     }
 
     public void monthlyAt(String name, Job job, int dayOfMonth, LocalTime time) {
-        state.scheduler.addTrigger(new MonthlyTrigger(name, job, dayOfMonth, time, state.zoneId));
+        trigger(name, job, new MonthlyTrigger(dayOfMonth, time));
+    }
+
+    public void trigger(String name, Job job, Trigger trigger) {
+        state.scheduler.addTriggerTask(name, job, trigger);
         state.triggerAdded = true;
     }
 
@@ -76,6 +71,5 @@ public final class SchedulerConfig {
     public static class State {
         Scheduler scheduler;
         boolean triggerAdded;
-        ZoneId zoneId = ZoneId.systemDefault();
     }
 }