+++ core-ng/src/main/java/core/framework/scheduler/Trigger.java
@@ -1,0 +1,11 @@
+package core.framework.scheduler;
+
+import java.time.ZonedDateTime;
+
+/**
+ * @author neo
+ */
+@FunctionalInterface
+public interface Trigger {
+    ZonedDateTime next(ZonedDateTime previous);
+}